Toda la funcionalidad de comentarios funcionando con validacion y tests

+++ src/main/java/org/springframework/samples/petclinic/web/ComentarioController.java
@@ -3,6 +3,7 @@ package org.springframework.samples.petclinic.web;
 import java.util.Collection;
 import java.util.List;
 
+import javax.validation.Valid;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,12 +11,15 @@ import org.springframework.samples.petclinic.model.Cliente;
 import org.springframework.samples.petclinic.model.Comentario;
 import org.springframework.samples.petclinic.model.Merchandasing;
 import org.springframework.samples.petclinic.model.Pelicula;
+import org.springframework.samples.petclinic.model.User;
 import org.springframework.samples.petclinic.model.Videojuego;
 import org.springframework.samples.petclinic.service.ClienteService;
 import org.springframework.samples.petclinic.service.ComentarioService;
 import org.springframework.samples.petclinic.service.MerchandasingService;
 import org.springframework.samples.petclinic.service.PeliculaService;
+import org.springframework.samples.petclinic.service.UserService;
 import org.springframework.samples.petclinic.service.VideojuegoService;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
@@ -48,6 +52,9 @@ public class ComentarioController {
 	@Autowired 
 	private final MerchandasingService merchandasingService;
 	
+	@Autowired
+	private final UserService userService;
+	
 	
 	private static final String VIEWS_COMENTARIOS_CREATE_OR_UPDATE_FORM_P = "comentarios/createOrUpdateComentarioForm";
 	
@@ -68,14 +75,40 @@ public class ComentarioController {
 		}
 		return vista;
 	}
+	//EN PROCESO
 	
+	@GetMapping(value = "pelicula/{peliculaId}")
+	public String listComentariosPelicula(@PathVariable("peliculaId") int peliculaId, ModelMap modelMap) {
+		String vista = "peliculas/comentariosProductoList";
+		List<Comentario> comentarios = comentarioService.findComentariosByPeliculaId(peliculaId);
+		modelMap.addAttribute("comentarios", comentarios);
+		return vista;
+	}
+	
+	@GetMapping(value = "videojuego/{videojuegoId}")
+	public String listComentariosVideojuego(@PathVariable("videojuegoId") int videojuegoId, ModelMap modelMap) {
+		String vista = "comentarios/comentariosProductoList";
+		List<Comentario> comentarios = comentarioService.findComentariosByVideojuegoId(videojuegoId);
+		modelMap.addAttribute("comentarios", comentarios);
+		return vista;
+	}
+	
+	@GetMapping(value = "merchandasing/{merchandasingId}")
+	public String listComentariosMerchandasing(@PathVariable("merchandasingId") int merchandasingId, ModelMap modelMap) {
+		String vista = "comentarios/comentariosProductoList";
+		List<Comentario> comentarios = comentarioService.findComentariosByMerchandasingId(merchandasingId);
+		modelMap.addAttribute("comentarios", comentarios);
+		return vista;
+	}
+
 	@Autowired
-	public ComentarioController(ComentarioService comentarioService, ClienteService clienteService, PeliculaService peliculaService, VideojuegoService videojuegoService, MerchandasingService merchandasingService) {
+	public ComentarioController(ComentarioService comentarioService, ClienteService clienteService, PeliculaService peliculaService, VideojuegoService videojuegoService, MerchandasingService merchandasingService, UserService userService) {
 		this.comentarioService = comentarioService;
 		this.clienteService = clienteService;
 		this.peliculaService = peliculaService;
 		this.videojuegoService = videojuegoService;
 		this.merchandasingService = merchandasingService;
+		this.userService = userService;
 	}
 	
 	@ModelAttribute("cliente")
@@ -104,8 +137,13 @@ public class ComentarioController {
 	}
 	
 	@GetMapping(value = "/pelicula/{peliculaId}/new")
-	public String initCreationFormPelicula(@PathVariable("peliculaId") int peliculaId, Cliente cliente, ModelMap model) {
+	public String initCreationFormComentarioPelicula(@PathVariable("peliculaId") int peliculaId, ModelMap model) {
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetails = (UserDetails) auth.getPrincipal();
+		User usuario = this.userService.findUser(userDetails.getUsername()).get();
 		
+		
+		Cliente cliente = clienteService.findClienteByUserName(usuario.getUsername());
 		Comentario comentario = new Comentario(cliente);
 		Pelicula pelicula = peliculaService.findPeliculaById(peliculaId);
 		
@@ -118,8 +156,13 @@ public class ComentarioController {
 	}
 	
 	@GetMapping(value = "/videojuego/{videojuegoId}/new")
-	public String initCreationFormVideojuego(@PathVariable("videojuegoId") int videojuegoId, Cliente cliente, ModelMap model) {
+	public String initCreationFormComentarioVideojuego(@PathVariable("videojuegoId") int videojuegoId, ModelMap model) {
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetails = (UserDetails) auth.getPrincipal();
+		User usuario = this.userService.findUser(userDetails.getUsername()).get();
 		
+		
+		Cliente cliente = clienteService.findClienteByUserName(usuario.getUsername());
 		Comentario comentario = new Comentario(cliente);
 		Videojuego videojuego = videojuegoService.findVideojuegoById(videojuegoId);
 		
@@ -131,8 +174,13 @@ public class ComentarioController {
 	}
 	
 	@GetMapping(value = "merchandasing/{merchandasingId}/new")
-	public String initCreationFormMerchandasing(@PathVariable("merchandasingId") int merchandasingId, Cliente cliente, ModelMap model) {
+	public String initCreationFormComentarioMerchandasing(@PathVariable("merchandasingId") int merchandasingId, ModelMap model) {
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetails = (UserDetails) auth.getPrincipal();
+		User usuario = this.userService.findUser(userDetails.getUsername()).get();
 		
+		
+		Cliente cliente = clienteService.findClienteByUserName(usuario.getUsername());
 		Comentario comentario = new Comentario(cliente);
 		Merchandasing merchandasing = this.merchandasingService.findMerchandasingById(merchandasingId);
 		
@@ -145,7 +193,7 @@ public class ComentarioController {
 	
 	
 	@PostMapping(value = "/pelicula/{peliculaId}/new")
-	public String processCreationFormPelicula(Comentario comentario, BindingResult result, ModelMap model) {	
+	public String processCreationFormComentarioPelicula(@Valid Comentario comentario, BindingResult result, ModelMap model) {	
 		if (result.hasErrors()) {
 			System.out.println(result.getAllErrors());
 			model.addAttribute("comentario", comentario);
@@ -181,7 +229,7 @@ public class ComentarioController {
 	}
 	
 	@PostMapping(value = "/videojuego/{videojuegoId}/new")
-	public String processCreationFormVideojuego(Comentario comentario, BindingResult result, ModelMap model) {
+	public String processCreationFormComentarioVideojuego(@Valid Comentario comentario, BindingResult result, ModelMap model) {
 		if(result.hasErrors()) {
 			System.out.println(result.getAllErrors());
 			model.addAttribute("comentario", comentario);
@@ -214,7 +262,7 @@ public class ComentarioController {
 	}
 	
 	@PostMapping(value = "/merchandasing/{merchandasingId}/new")
-	public String processCreationFormMerchandasing(Comentario comentario , BindingResult result, ModelMap model) {
+	public String processCreationFormComentarioMerchandasing(@Valid Comentario comentario , BindingResult result, ModelMap model) {
 		if(result.hasErrors()) {
 			model.addAttribute("comentario", comentario);
 			model.addAttribute("message", "El comentario no se ha podido crear");
@@ -245,7 +293,7 @@ public class ComentarioController {
 	}
 	
 	@GetMapping(value = "/comentario/{comentarioId}/edit")
-	public String initUpdateFormComentarioPelicula(@PathVariable("comentarioId") int comentarioId, ModelMap model) {
+	public String initUpdateFormComentario(@PathVariable("comentarioId") int comentarioId, ModelMap model) {
 		String res = "";
 		Comentario comentario = comentarioService.findCommentById(comentarioId);
 		model.put("comentario", comentario);
@@ -260,7 +308,7 @@ public class ComentarioController {
 	}
 	
 	@PostMapping(value = "/comentario/{comentarioId}/edit")
-	public String processUpdateFormComentarioPelicula(Comentario comentario, BindingResult result, Cliente cliente,@PathVariable("comentarioId") int comentarioId, ModelMap model) {
+	public String processUpdateFormComentario(@Valid Comentario comentario, BindingResult result, Cliente cliente,@PathVariable("comentarioId") int comentarioId, ModelMap model) {
 		if (result.hasErrors()) {
 			model.put("comentario", comentario);
 			return VIEWS_COMENTARIOS_CREATE_OR_UPDATE_FORM_P;
@@ -277,13 +325,13 @@ public class ComentarioController {
 	@GetMapping("comentario/{comentarioId}/delete")
 	public String deleteComentario(@PathVariable("comentarioId") int comentarioId, ModelMap model) {
 		Comentario comentario = comentarioService.findCommentById(comentarioId);
-		String view = "comentarios/comentariosList";
-		if(comentario!=null) {
+		Cliente cliente = comentario.getCliente();
+		if(comentario!=null && cliente!=null) {
 			comentarioService.deleteComment(comentario);
 		} else {
 			model.addAttribute("message", "ERROR!");
 		}
-		return view;
+		return "redirect:/comentarios/" + cliente.getId();
 	}
